Adds the function "toFile()" used to save the CFDI in a file.
Now is possible to save the CFDI XML content directly to a file by
calling this function and passing as parameter a file path. If file
exists an exception will be thrown unless a "true" is given as a second
parameter to the function, in such case file will be overwritten.